fix(type,tooling): assert certificate not None for mypy, move dispatch tests out of math ownership
